deps: update typescript 4
