+++ package.json
@@ -14,50 +14,50 @@
     "url": "https://opencollective.com/verdaccio"
   },
   "devDependencies": {
-    "@babel/cli": "7.11.6",
-    "@babel/core": "7.11.6",
-    "@babel/node": "7.10.5",
-    "@babel/plugin-proposal-class-properties": "7.10.4",
-    "@babel/plugin-proposal-decorators": "7.10.5",
-    "@babel/plugin-proposal-export-namespace-from": "7.10.4",
-    "@babel/plugin-proposal-function-sent": "7.10.4",
-    "@babel/plugin-proposal-json-strings": "7.10.4",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "7.10.4",
-    "@babel/plugin-proposal-numeric-separator": "7.10.4",
-    "@babel/plugin-proposal-object-rest-spread": "7.11.0",
-    "@babel/plugin-proposal-optional-chaining": "7.11.0",
-    "@babel/plugin-proposal-throw-expressions": "7.10.4",
+    "@babel/cli": "7.12.1",
+    "@babel/core": "7.12.3",
+    "@babel/node": "7.12.6",
+    "@babel/plugin-proposal-class-properties": "7.12.1",
+    "@babel/plugin-proposal-decorators": "7.12.1",
+    "@babel/plugin-proposal-export-namespace-from": "7.12.1",
+    "@babel/plugin-proposal-function-sent": "7.12.1",
+    "@babel/plugin-proposal-json-strings": "7.12.1",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "7.12.1",
+    "@babel/plugin-proposal-numeric-separator": "7.12.5",
+    "@babel/plugin-proposal-object-rest-spread": "7.12.1",
+    "@babel/plugin-proposal-optional-chaining": "7.12.1",
+    "@babel/plugin-proposal-throw-expressions": "7.12.1",
     "@babel/plugin-syntax-dynamic-import": "7.8.3",
     "@babel/plugin-syntax-import-meta": "7.10.4",
-    "@babel/plugin-transform-async-to-generator": "7.10.4",
-    "@babel/plugin-transform-classes": "7.10.4",
-    "@babel/plugin-transform-runtime": "7.11.5",
-    "@babel/polyfill": "7.11.5",
-    "@babel/preset-env": "7.11.5",
-    "@babel/preset-react": "7.10.4",
-    "@babel/preset-typescript": "7.10.4",
-    "@babel/register": "7.11.5",
-    "@babel/runtime": "7.11.2",
-    "@changesets/changelog-github": "0.1.1",
-    "@changesets/cli": "2.10.1",
+    "@babel/plugin-transform-async-to-generator": "7.12.1",
+    "@babel/plugin-transform-classes": "7.12.1",
+    "@babel/plugin-transform-runtime": "7.12.1",
+    "@babel/polyfill": "7.12.1",
+    "@babel/preset-env": "7.12.1",
+    "@babel/preset-react": "7.12.5",
+    "@babel/preset-typescript": "7.12.1",
+    "@babel/register": "7.12.1",
+    "@babel/runtime": "7.12.5",
+    "@changesets/changelog-github": "0.2.7",
+    "@changesets/cli": "2.11.2",
     "@commitlint/cli": "8.3.5",
     "@commitlint/config-conventional": "8.2.0",
     "@octokit/rest": "17.0.0",
     "@types/async": "3.2.3",
     "@types/express": "4.17.6",
     "@types/http-errors": "1.6.3",
-    "@types/jest": "26.0.1",
-    "@types/lodash": "4.14.156",
+    "@types/jest": "^26.0.15",
+    "@types/lodash": "4.14.165",
     "@types/mime": "2.0.2",
     "@types/minimatch": "3.0.3",
-    "@types/node": "14.0.14",
+    "@types/node": "^14.14.7",
     "@types/request": "2.48.3",
     "@types/semver": "7.2.0",
     "@types/supertest": "2.0.9",
-    "@typescript-eslint/eslint-plugin": "3.10.1",
-    "@typescript-eslint/parser": "3.10.1",
+    "@typescript-eslint/eslint-plugin": "4.7.0",
+    "@typescript-eslint/parser": "4.7.0",
     "@verdaccio/types": "workspace:*",
-    "@verdaccio/ui-theme": "latest",
+    "@verdaccio/ui-theme": "^1.14.0",
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "10.1.0",
     "babel-jest": "26.1.0",
@@ -95,7 +95,7 @@
     "standard-version": "8.0.0",
     "supertest": "4.0.2",
     "ts-node": "^9.0.0",
-    "typescript": "^3.9.7",
+    "typescript": "^4.0.5",
     "verdaccio": "workspace:*",
     "verdaccio-audit": "workspace:*",
     "verdaccio-auth-memory": "workspace:*",
