+++ packages/core/local-storage/src/local-fs.ts
@@ -1,3 +1,5 @@
+/* eslint-disable no-undef */
+
 import fs from 'fs';
 import path from 'path';
 import buildDebug from 'debug';
