+++ packages/mock/src/server.ts
@@ -4,9 +4,8 @@ import buildDebug from 'debug';
 
 import { API_MESSAGE, HEADERS, HTTP_STATUS, TOKEN_BASIC } from '@verdaccio/dev-commons';
 import { buildToken } from '@verdaccio/utils';
-import smartRequest from './request';
+import smartRequest, { PromiseAssert } from './request';
 
-import { IServerBridge } from './types';
 import { CREDENTIALS } from './constants';
 import getPackage from './fixtures/package';
 
@@ -15,6 +14,44 @@ const buildAuthHeader = (user, pass): string => {
 };
 
 const debug = buildDebug('verdaccio:mock:server');
+
+export interface IRequestPromise {
+  status(reason: any): any;
+  body_ok(reason: any): any;
+  body_error(reason: any): any;
+  request(reason: any): any;
+  response(reason: any): any;
+  send(reason: any): any;
+}
+
+export interface IServerBridge {
+  url: string;
+  userAgent: string;
+  authstr: string;
+  request(options: any): typeof PromiseAssert;
+  auth(name: string, password: string): IRequestPromise;
+  auth(name: string, password: string): IRequestPromise;
+  logout(token: string): Promise<any>;
+  getPackage(name: string): Promise<any>;
+  putPackage(name: string, data: any): Promise<any>;
+  putVersion(name: string, version: string, data: any): Promise<any>;
+  getTarball(name: string, filename: string): Promise<any>;
+  putTarball(name: string, filename: string, data: any): Promise<any>;
+  removeTarball(name: string): Promise<any>;
+  removeSingleTarball(name: string, filename: string): Promise<any>;
+  addTag(name: string, tag: string, version: string): Promise<any>;
+  putTarballIncomplete(
+    name: string,
+    filename: string,
+    data: any,
+    size: number,
+    cb: Function
+  ): Promise<any>;
+  addPackage(name: string): Promise<any>;
+  whoami(): Promise<any>;
+  ping(): Promise<any>;
+  debug(): IRequestPromise;
+}
 
 export default class Server implements IServerBridge {
   public url: string;
