+++ packages/verdaccio/package.json
@@ -36,20 +36,20 @@
   },
   "homepage": "https://verdaccio.org",
   "dependencies": {
+    "@verdaccio/cli": "workspace:5.0.0-alpha.0",
     "@verdaccio/hooks": "workspace:5.0.0-alpha.0",
-    "@verdaccio/utils": "workspace:5.0.0-alpha.0",
     "@verdaccio/logger": "workspace:5.0.0-alpha.0",
     "@verdaccio/mock": "workspace:5.0.0-alpha.0",
     "@verdaccio/node-api": "workspace:5.0.0-alpha.0",
-    "@verdaccio/cli": "workspace:5.0.0-alpha.0",
-    "verdaccio-htpasswd": "9.7.2",
-    "@verdaccio/ui-theme": "^1.12.1"
+    "@verdaccio/ui-theme": "^1.14.0",
+    "@verdaccio/utils": "workspace:5.0.0-alpha.0",
+    "verdaccio-htpasswd": "9.7.2"
   },
   "devDependencies": {
-    "@verdaccio/config": "workspace:5.0.0-alpha.0",
     "@verdaccio/auth": "workspace:5.0.0-alpha.0",
-    "@verdaccio/store": "workspace:5.0.0-alpha.0",
-    "@verdaccio/dev-commons": "workspace:*"
+    "@verdaccio/config": "workspace:5.0.0-alpha.0",
+    "@verdaccio/dev-commons": "workspace:*",
+    "@verdaccio/store": "workspace:5.0.0-alpha.0"
   },
   "keywords": [
     "private",
