+++ website/.eslintrc
@@ -4,7 +4,7 @@
   },
   "settings": {
     "react": {
-      "version": "detect"
+      "version": "16"
     }
   },
   "rules": {
